개인루틴패키지 도메인에 포함
