+++ src/main/java/com/goormi/routine/domain/userActivity/service/UserActivityServiceImpl.java
@@ -11,8 +11,8 @@ import com.goormi.routine.domain.userActivity.dto.UserActivityResponse;
 import com.goormi.routine.domain.userActivity.entity.ActivityType;
 import com.goormi.routine.domain.userActivity.entity.UserActivity;
 import com.goormi.routine.domain.userActivity.repository.UserActivityRepository;
-import com.goormi.routine.personal_routines.domain.PersonalRoutine;
-import com.goormi.routine.personal_routines.repository.PersonalRoutineRepository;
+import com.goormi.routine.domain.personal_routines.domain.PersonalRoutine;
+import com.goormi.routine.domain.personal_routines.repository.PersonalRoutineRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
