+++ src/test/java/com/goormi/routine/domain/group/service/UserActivityServiceTest.java
@@ -11,11 +11,11 @@ import com.goormi.routine.domain.userActivity.dto.UserActivityRequest;
 import com.goormi.routine.domain.userActivity.dto.UserActivityResponse;
 import com.goormi.routine.domain.userActivity.entity.ActivityType;
 import com.goormi.routine.domain.userActivity.service.UserActivityService;
-import com.goormi.routine.personal_routines.domain.PersonalRoutine;
-import com.goormi.routine.personal_routines.dto.PersonalRoutineRequest;
-import com.goormi.routine.personal_routines.dto.PersonalRoutineResponse;
-import com.goormi.routine.personal_routines.repository.PersonalRoutineRepository;
-import com.goormi.routine.personal_routines.service.PersonalRoutineService;
+import com.goormi.routine.domain.personal_routines.domain.PersonalRoutine;
+import com.goormi.routine.domain.personal_routines.dto.PersonalRoutineRequest;
+import com.goormi.routine.domain.personal_routines.dto.PersonalRoutineResponse;
+import com.goormi.routine.domain.personal_routines.repository.PersonalRoutineRepository;
+import com.goormi.routine.domain.personal_routines.service.PersonalRoutineService;
 import jakarta.transaction.Transactional;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
